Move payload parsing into payload specific classes
